SetThreadname Changes requested after review
