+++ base/src/PipeLine.cpp
@@ -1,66 +1,13 @@
 #include "stdafx.h"
 #include "PipeLine.h"
 #include "Module.h"
+#include "Utils.h"
 #include <string>
-#ifndef _WIN64
-#include <pthread.h>
-#endif
-
-#ifdef _WIN64
-	// Windows
-	const DWORD MS_VC_EXCEPTION = 0x406D1388;
-	#pragma pack(push,8)
-	typedef struct THREADNAME_INFO {
-		DWORD dwType; // Must be 0x1000.
-		LPCSTR szName; // Pointer to name (in user addr space).
-		DWORD dwThreadID; // Thread ID (-1=caller thread).
-		DWORD dwFlags;
-	} THREADNAME_INFO;
-	#pragma pack(pop)
-
-	void _SetThreadNameWIN(DWORD threadID, const char* threadName) {
-		THREADNAME_INFO info;
-		info.dwType = 0x1000;
-		info.szName = threadName;
-		info.dwThreadID = threadID;
-		info.dwFlags = 0;
-		__try
-		{
-			RaiseException(MS_VC_EXCEPTION, 0, sizeof(info) / sizeof(ULONG_PTR), (ULONG_PTR *)&info);
-		}
-		__except (EXCEPTION_EXECUTE_HANDLER)
-		{
-		}
-	}
-	void SetThreadNameWIN(boost::thread::id threadId, std::string threadName)
-	{
-		// convert string to char*
-		const char* cchar = threadName.c_str();
-		// convert HEX string to DWORD
-		unsigned int dwThreadId;
-		std::stringstream ss;
-		ss << std::hex << threadId;
-		ss >> dwThreadId;
-		// set thread name
-		_SetThreadNameWIN((DWORD)dwThreadId, cchar);
-	}
-#endif
-
-void setModuleThreadName(boost::thread &thread, std::string moduleID)
-{
-#ifdef _WIN64
-	SetThreadNameWIN((thread).get_id(), moduleID);
-#else
-	auto ptr = thread.native_handle();
-	pthread_setname_np(ptr, moduleID.c_str());
-#endif
-}
 
 PipeLine::~PipeLine()
 {
 	//LOG_INFO << "Dest'r ~PipeLine" << mName << endl;
 }
-
 
 bool PipeLine::appendModule(boost::shared_ptr<Module> pModule)
 {
@@ -200,7 +147,7 @@ void PipeLine::run_all_threaded()
 	{
 		Module& m = *(i->get());
 		m.myThread = boost::thread(ref(m));
-		setModuleThreadName(m.myThread, m.getId());
+		Utils::setModuleThreadName(m.myThread, m.getId());
 	}
 	mPlay = true;
 }
